requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
dj-database-url==3.0.1
Django==4.2.23
django-grappelli==4.0.2
django-import-export==4.3.9
python-decouple==3.8
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
